refactor: organizing files nestjs style and database connection on its own file
